add: RestUtils method; fix: OAuth2
