+++ src/main/java/de/wirvsvirus/neighborhoodaid/api/stats/HealthEndpoint.java
@@ -1,6 +1,7 @@
 package de.wirvsvirus.neighborhoodaid.api.stats;
 
 import de.wirvsvirus.neighborhoodaid.api.Endpoint;
+import de.wirvsvirus.neighborhoodaid.api.utils.RestUtils;
 import io.vertx.core.Vertx;
 import io.vertx.ext.web.Router;
 import org.apache.commons.lang3.time.DurationFormatUtils;
@@ -15,7 +16,7 @@ public class HealthEndpoint implements Endpoint {
         router.get().handler(ctx -> {
             final var uptime = ManagementFactory.getRuntimeMXBean().getUptime();
             final var formatted = DurationFormatUtils.formatDurationHMS(uptime);
-            ctx.response().end("<h1 style=\"color:green\")>Service is running for " + formatted + "</h1>");
+            RestUtils.endResponseWithHtmlSuccess(ctx, 200, "Service is running for " + formatted);
         });
     }
 }
