--- src/main/java/de/wirvsvirus/neighborhoodaid/oauth/GCallback.java
@@ -1,14 +1,0 @@
-package de.wirvsvirus.neighborhoodaid.oauth;
-
-import de.wirvsvirus.neighborhoodaid.api.Endpoint;
-import io.vertx.core.Vertx;
-import io.vertx.ext.web.Router;
-import org.jetbrains.annotations.NotNull;
-
-public class GCallback implements Endpoint {
-
-    @Override
-    public void setupRouting(@NotNull Vertx vertx, @NotNull Router router) {
-        router.r
-    }
-}
